Upgrade core, dependencies, base schema, gradle

+++ src/main/java/org/veupathdb/service/eda/access/controller/EndUserController.java
@@ -27,8 +27,8 @@ public class EndUserController implements DatasetEndUsers
   @Override
   public GetDatasetEndUsersResponse getDatasetEndUsers(
     final String datasetId,
-    final int limit,
-    final int offset,
+    final Long limit,
+    final Long offset,
     final ApprovalStatus approval
   ) {
     return GetDatasetEndUsersResponse.respond200WithApplicationJson(
